alterando o processo de gravacao

+++ src/main/java/persistencia/ArtistaDAO.java
@@ -8,23 +8,24 @@ import java.util.ArrayList;
 import java.util.List;
 
 import dominio.Artista;
+import erro.Excessao;
 
 public class ArtistaDAO {
-	
+
 	private Connection con = Conexao.getConnection();
 
 	public void cadastrar(Artista artista) {
 
 		String sql = "INSERT INTO TB_ARTISTA (cache,nacionalidade,nascimento,nome) VALUES (?,?,?,?)";
-	
+
 		PreparedStatement preparador;
 		try {
 			preparador = con.prepareStatement(sql);
 			preparador.setBigDecimal(1, artista.getCache());
 			preparador.setString(2, artista.getNacionalidade());
-			
+
 			java.sql.Date mySqlDate = new java.sql.Date(artista.getNascimento().getTime());
-		
+
 			preparador.setDate(3, mySqlDate);
 			preparador.setString(4, artista.getNome());
 			preparador.execute();
@@ -36,19 +37,18 @@ public class ArtistaDAO {
 		}
 
 	}
-	
-	
-	
-	
-	public void salvar(Artista artista) {
-		if(artista.getCodArtista()!=null && artista.getCodArtista()!=0) {
+
+	public void salvar(Artista artista) throws Excessao {
+		if (artista.getCodArtista() != null && artista.getCodArtista() != 0) {
 			this.alterar(artista);
 		} else {
+			Artista artistaExiste = buscarNomeExato(artista.getNome());
+			if (artistaExiste != null) {
+				throw new Excessao("Ja existe um artista com este nome!", 1);
+			}
 			this.cadastrar(artista);
 		}
 	}
-	
-	
 
 	public void alterar(Artista artista) {
 
@@ -59,9 +59,9 @@ public class ArtistaDAO {
 			preparador = con.prepareStatement(sql);
 			preparador.setBigDecimal(1, artista.getCache());
 			preparador.setString(2, artista.getNacionalidade());
-			
+
 			java.sql.Date mySqlDate = new java.sql.Date(artista.getNascimento().getTime());
-			
+
 			preparador.setDate(3, mySqlDate);
 			preparador.setString(4, artista.getNome());
 			preparador.setLong(5, artista.getCodArtista());
@@ -75,7 +75,11 @@ public class ArtistaDAO {
 
 	}
 
-	public void excluir(Artista artista) {
+	public void excluir(Artista artista) throws Excessao {
+
+		if (!artista.getParticipacoes().isEmpty()) {
+			throw new Excessao("Exclusão não permitida pois este artista tem participações!", 2);
+		}
 
 		String sql = "DELETE FROM TB_ARTISTA WHERE codArtista=?";
 
@@ -96,6 +100,35 @@ public class ArtistaDAO {
 	public List<Artista> buscarTodos() {
 
 		String sql = "SELECT * FROM TB_ARTISTA";
+
+		List<Artista> artistas = new ArrayList<Artista>();
+
+		PreparedStatement preparador;
+		try {
+			preparador = con.prepareStatement(sql);
+
+			ResultSet resultado = preparador.executeQuery();
+			while (resultado.next()) {
+				Artista artista = new Artista();
+				artista.setCodArtista(resultado.getLong("codArtista"));
+				artista.setCache(resultado.getBigDecimal("cache"));
+				artista.setNacionalidade(resultado.getString("nacionalidade"));
+				artista.setNascimento(resultado.getDate("nascimento"));
+				artista.setNome(resultado.getString("nome"));
+				artistas.add(artista);
+			}
+			preparador.close();
+			System.out.println("Listado com sucesso");
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+		return artistas;
+
+	}
+
+	public List<Artista> buscarTodosOrdenadosPorNome() {
+
+		String sql = "SELECT * FROM TB_ARTISTA ORDER BY nome";
 
 		List<Artista> artistas = new ArrayList<Artista>();
 
@@ -148,15 +181,41 @@ public class ArtistaDAO {
 		return artista;
 
 	}
-	
-	
+
+	public Artista buscarNomeExato(String nome) {
+		String sql = "SELECT * FROM TB_ARTISTA WHERE nome =?";
+		PreparedStatement preparador;
+		Artista artista = null;
+		try {
+			preparador = con.prepareStatement(sql);
+			preparador.setString(1, nome);
+			ResultSet resultado = preparador.executeQuery();
+
+			if (resultado.next()) {
+				artista = new Artista();
+				artista.setCodArtista(resultado.getLong("codArtista"));
+				artista.setCache(resultado.getBigDecimal("cache"));
+				artista.setNacionalidade(resultado.getString("nacionalidade"));
+				artista.setNascimento(resultado.getDate("nascimento"));
+				artista.setNome(resultado.getString("nome"));
+			}
+
+			preparador.close();
+
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+		return artista;
+
+	}
+
 	public List<Artista> buscarPorNome(String nome) {
 		String sql = "SELECT * FROM TB_ARTISTA WHERE nome LIKE ?";
 		PreparedStatement preparador;
 		List<Artista> artistas = new ArrayList<Artista>();
 		try {
 			preparador = con.prepareStatement(sql);
-			preparador.setString(1, "%"+nome+"%");
+			preparador.setString(1, "%" + nome + "%");
 			ResultSet resultado = preparador.executeQuery();
 
 			while (resultado.next()) {
@@ -177,8 +236,5 @@ public class ArtistaDAO {
 		return artistas;
 
 	}
-	
-	
-	
 
 }
